+++ src/Coordination/tests/gtest_coordination_common.h
@@ -11,6 +11,8 @@
 #include <Coordination/KeeperStorage_fwd.h>
 #include <Coordination/KeeperCommon.h>
 
+#include <Common/ZooKeeper/ZooKeeperCommon.h>
+
 #include <Disks/DiskLocal.h>
 
 #include <Poco/Logger.h>
@@ -118,9 +120,9 @@ void addNode(Storage & storage, const std::string & path, const std::string & da
     node.acl_id = acl_id;
     storage.container.insertOrReplace(path, node);
     auto child_it = storage.container.find(path);
-    auto child_path = DB::getBaseNodeName(child_it->key);
+    auto child_path = Coordination::getBaseNodeName(child_it->key);
     storage.container.updateValue(
-        DB::parentNodePath(StringRef{path}),
+        Coordination::parentNodePath(StringRef{path}),
         [&](auto & parent)
         {
             parent.addChild(child_path);
